auto add & push

+++ src/parsing_lightsandcam.c
@@ -35,7 +35,7 @@ void	parse_camera(t_data *d)
 	d->camera.point = parse_tpoint(d, d->parse.elmnts[1], E_CAMERA1);
 	d->camera.vector = parse_tvector(d, d->parse.elmnts[2], E_CAMERA2);
 	d->parse.min = 0;
-	d->parse.max = 180;
+	d->parse.max = (double)180;
 	d->camera.fieldofview = parse_int(d, d->parse.elmnts[3], E_CAMERA3);
 	d->fov_ratio = tan(d->camera.fieldofview / 2 * M_PI / 180);
 }
